+++ ctview/map_class/raster_generation.py
@@ -2,25 +2,18 @@ import logging as log
 import os
 import tempfile
 from collections.abc import Iterable
-from typing import Tuple
 
-import hydra
 import numpy as np
 from omegaconf import DictConfig
-from osgeo import gdal
 
-from ctview import clip_raster, map_DXM, utils_pdal, utils_raster
+from ctview import map_DXM, utils_raster
+from ctview.add_color import convert_raster_with_color_metadata_to_rgb
 from ctview.map_class.classes_mapping import (
     check_and_list_original_classes_to_keep,
     compute_binary_class,
     convert_class_array_to_precedence_array,
-    create_class_raster_raw_deprecated,
 )
-from ctview.map_class.post_processing import (
-    add_color_to_raster,
-    fill_gaps_raster,
-    post_processing,
-)
+from ctview.map_class.post_processing import post_processing
 
 
 def generate_class_raster_raw(
@@ -186,11 +179,7 @@ def generate_class_raster(
             priorities=config_class.precedence_classes,
         )
 
-        post_processed_class_map = post_processing(
-            flatten_array,
-            config_class.post_processing.smoothing.nconnectedness,
-            config_class.post_processing.smoothing.threshold,
-        )
+        post_processed_class_map = post_processing(flatten_array, config_class.post_processing)
 
         utils_raster.write_single_band_raster_to_file(
             input_array=post_processed_class_map,
@@ -244,10 +233,9 @@ def generate_pretty_class_raster_from_single_band_raster(
         dxm_raw_tmp_file = os.path.join(tmpdir, f"{tilename}_dxm_raw{ext}")
         dxm_hillshade_tmp_file = os.path.join(tmpdir, f"{tilename}_dxm_hillshade{ext}")
 
-        ds = gdal.Open(input_raster)
-        ds = gdal.Translate(colored_tmp_file, ds, rgbExpand="rgb")  # Use colors in metadata
-        ds = None  # close file
-        output_raster = os.path.join(output_dir, f"{tilename}_class_pretty{ext}")
+        convert_raster_with_color_metadata_to_rgb(input_raster, colored_tmp_file)
+
+        output_raster = os.path.join(output_dir, f"{tilename}{ext}")
         map_DXM.add_dxm_hillshade_to_raster(
             input_raster=colored_tmp_file,
             input_pointcloud=str(input_las),
@@ -260,174 +248,3 @@ def generate_pretty_class_raster_from_single_band_raster(
             hillshade_calc=config_class.hillshade_calc,
             config_io=config_io,
         )
-
-
-def create_map_class_raster_with_postprocessing_color_and_hillshade(
-    input_las: str,
-    tilename: str,
-    config_class: DictConfig | dict,
-    config_io: DictConfig | dict,
-    output_bounds: Tuple,
-):
-    """Create a raster of class with post processing steps:
-    * fill gaps with a method of gdal
-    * colorisation
-    * mix with a hillshade that comes from an elevation model
-
-    Args:
-        input_las (str): path to the input las file
-        tilename (str): tilename used to generate the output filename
-        config_class (DictConfig | dict):  hydra configuration with the class map parameters
-        eg. {
-          pixel_size: 0.5  # pixel size of the output raster
-          dxm_filter:  # filter used to generate dsm, the dsm is used to generate hillshade
-              dimension: Classification
-              keep_values: [1, 2, 3, 4, 5, 6, 9, 17, 64, 65, 66, 67, 202]
-          lut_filename: LUT_CLASS.txt  # filename for the lut that colorizes the classification map
-          output_subdir: CC_6_fusion  # folder name for the final class map output
-          hillshade_calc: "254*((A*(0.5*(B/255)+0.25))>254)+(A*(0.5*(B/255)+0.25))*((A*(0.5*(B/255)+0.25))<=254)"
-          # formula used to mix the classification map with its hillshade
-          intermediate_dirs:  # paths to the saved intermediate results
-              CC_raw: CC_1_raw
-              CC_raw_color: CC_2_bcolor
-              CC_fillgap: CC_3_fg
-              CC_fillgap_color: CC_4_fgcolor
-              CC_crop: CC_5_crop
-              dxm_raw: "DSM"
-              dxm_hillshade: "tmp_dsm/hillshade"
-
-        }
-        config_io (DictConfig | dict): hydra configuration with the general io parameters
-        eg.  {
-            input_filename: null
-            input_dir: null
-            output_dir: null
-            spatial_reference: EPSG:2154
-            lut_folder: LUT
-            extension: .tif
-            raster_driver: GTiff
-            no_data_value: -9999
-            tile_geometry:
-                tile_coord_scale: 1000
-                tile_width: 1000
-            }
-        output_bounds (Tuple): bounds of the output raster file ([minx,maxx],[miny, maxy]).
-            Should correspond to the las input file before adding any buffer
-    """
-    log.basicConfig(level=log.INFO, format="%(message)s")
-
-    # File names
-    input_las_name = os.path.basename(input_las)
-    in_points = utils_pdal.read_las_file(input_las)
-
-    log.info(f"\nMAP OF CLASS : file : {input_las_name}")
-
-    # Output_folder_names
-    out_dir = config_io.output_dir
-    inter_dirs = config_class.intermediate_dirs
-    ext = config_io.extension
-
-    with tempfile.TemporaryDirectory(prefix="tmp_class", dir="tmp") as tmpdir:
-        if inter_dirs.CC_raw:
-            raster_raw = os.path.join(out_dir, inter_dirs.CC_raw, f"{tilename}_raw{ext}")
-        else:
-            raster_raw = os.path.join(tmpdir, f"{tilename}_raw{ext}")
-        if inter_dirs.CC_raw_color:
-            raster_raw_color = os.path.join(out_dir, inter_dirs.CC_raw_color, f"{tilename}_raw_color{ext}")
-
-        if inter_dirs.CC_fillgap:
-            raster_fillgap = os.path.join(out_dir, inter_dirs.CC_fillgap, f"{tilename}_fillgap{ext}")
-        else:
-            raster_fillgap = os.path.join(tmpdir, f"{tilename}_fillgap{ext}")
-        if inter_dirs.CC_fillgap_color:
-            raster_fg_color = os.path.join(out_dir, inter_dirs.CC_fillgap_color, f"{tilename}_fillgap_color{ext}")
-        else:
-            raster_fg_color = os.path.join(tmpdir, f"{tilename}_fillgaps_color{ext}")
-        if inter_dirs.CC_crop:
-            raster_fg_color_clip = os.path.join(out_dir, inter_dirs.CC_crop, f"{tilename}_fillgap_color_clip{ext}")
-        else:
-            raster_fg_color_clip = os.path.join(tmpdir, f"{tilename}_fillgap_color_clip{ext}")
-
-        if inter_dirs.dxm_raw:
-            raster_class_map_dxm_raw = os.path.join(out_dir, inter_dirs.dxm_raw, f"{tilename}_interp{ext}")
-        else:
-            raster_class_map_dxm_raw = os.path.join(tmpdir, f"{tilename}_interp{ext}")
-        if inter_dirs.dxm_hillshade:
-            raster_class_map_dxm_hs = os.path.join(out_dir, inter_dirs.dxm_hillshade, f"{tilename}_hillshade{ext}")
-        else:
-            raster_class_map_dxm_hs = os.path.join(tmpdir, f"{tilename}_hillshade{ext}")
-
-        raster_class_map = os.path.join(out_dir, config_class.output_subdir, f"{tilename}_fusion_DSM_class{ext}")
-        os.makedirs(os.path.dirname(raster_class_map), exist_ok=True)
-
-        log.info("Step 1 : Write raster raw")
-        create_class_raster_raw_deprecated(
-            in_points=in_points,
-            output_file=raster_raw,
-            res=config_class.pixel_size,
-            raster_driver=config_io.raster_driver,
-            no_data_value=config_io.no_data_value,
-        )
-
-        if inter_dirs["CC_raw_color"]:
-            log.info("Step 1.5 : Write raster raw colorized")
-            add_color_to_raster(
-                in_raster=raster_raw,
-                output_file=raster_raw_color,
-                LUT=os.path.join(config_io.lut_folder, config_class.lut_filename),
-            )
-
-        log.info("Step 2 : Write raster with fillgap")
-        fill_gaps_raster(
-            in_raster=raster_raw,
-            output_file=raster_fillgap,
-            raster_driver=config_io.raster_driver,
-        )
-
-        log.info("Step 2 : Write raster with fillgap and color")
-        add_color_to_raster(
-            in_raster=raster_fillgap,
-            output_file=raster_fg_color,
-            LUT=os.path.join(config_io.lut_folder, config_class.lut_filename),
-        )
-        log.info("Step 3 : Write raster with fillgap, color + clipped at the right bounds")
-        clip_raster.clip_raster(
-            input_raster=raster_fg_color,
-            output_raster=raster_fg_color_clip,
-            bounds=output_bounds,
-            raster_driver=config_io.raster_driver,
-        )
-
-        map_DXM.add_dxm_hillshade_to_raster(
-            input_raster=raster_fg_color_clip,
-            input_pointcloud=str(input_las),
-            output_raster=raster_class_map,
-            pixel_size=config_class.pixel_size,
-            dxm_filter_dimension=config_class.dxm_filter.dimension,
-            dxm_filter_keep_values=config_class.dxm_filter.keep_values,
-            output_dxm_raw=raster_class_map_dxm_raw,
-            output_dxm_hillshade=raster_class_map_dxm_hs,
-            hillshade_calc=config_class.hillshade_calc,
-            config_io=config_io,
-        )
-
-
-@hydra.main(config_path="../../configs/", config_name="config_ctview.yaml", version_base="1.2")
-def main(config: DictConfig):
-    log.basicConfig(level=log.INFO, format="%(message)s")
-    initial_las_file = os.path.join(config.io.input_dir, config.io.input_filename)
-    las_bounds = utils_pdal.get_bounds_from_las(initial_las_file)
-    os.makedirs(config.io.output_dir, exist_ok=True)
-    create_map_class_raster_with_postprocessing_color_and_hillshade(
-        input_las=initial_las_file,
-        tilename=os.path.splitext(os.path.basename(initial_las_file))[0],
-        config_class=config.class_map,
-        config_io=config.io,
-        output_bounds=las_bounds,
-    )
-    log.info("END.")
-
-
-if __name__ == "__main__":
-    gdal.UseExceptions()
-    main()